Add test for `subcommand_create_tx()`
