+++ tests/fixtures/transaction_inputs.json
@@ -1,0 +1,51 @@
+[
+	{
+		"recipient": "ST2ZRX0K27GW0SP3GJCEMHD95TQGJMKB7G9Y0X1MH",
+		"fee": "300",
+		"amount": "10000",
+		"publicKeys": [
+      "02b30fafab3a12372c5d150d567034f37d60a91168009a779498168b0e9d8ec7f2",
+      "03ce61f1d155738a5e434fc8a61c3e104f891d1ec71576e8ad85abb68b34670d35",
+      "03ef2340518b5867b23598a9cf74611f8b98064f7d55cdb8c107c67b5efcbc5c77"
+		],
+		"numSignatures": 3,
+		"nonce": "4",
+		"network": "testnet"
+	},
+	{
+		"recipient": "ST2ZRX0K27GW0SP3GJCEMHD95TQGJMKB7G9Y0X1MH",
+		"fee": "777",
+		"amount": "100000",
+		"publicKeys": [
+      "02b30fafab3a12372c5d150d567034f37d60a91168009a779498168b0e9d8ec7f2",
+      "03ce61f1d155738a5e434fc8a61c3e104f891d1ec71576e8ad85abb68b34670d35",
+      "03ef2340518b5867b23598a9cf74611f8b98064f7d55cdb8c107c67b5efcbc5c77"
+		],
+		"numSignatures": 2,
+		"network": "testnet"
+	},
+	{
+		"recipient": "ST2ZRX0K27GW0SP3GJCEMHD95TQGJMKB7G9Y0X1MH",
+		"fee": "300",
+		"amount": "50000",
+		"publicKeys": [
+      "02b30fafab3a12372c5d150d567034f37d60a91168009a779498168b0e9d8ec7f2",
+      "03ce61f1d155738a5e434fc8a61c3e104f891d1ec71576e8ad85abb68b34670d35",
+      "03ef2340518b5867b23598a9cf74611f8b98064f7d55cdb8c107c67b5efcbc5c77"
+		],
+		"numSignatures": 1,
+		"nonce": "1"
+	},
+	{
+		"recipient": "ST2ZRX0K27GW0SP3GJCEMHD95TQGJMKB7G9Y0X1MH",
+		"fee": "777",
+		"amount": "100000",
+		"publicKeys": [
+      "02b30fafab3a12372c5d150d567034f37d60a91168009a779498168b0e9d8ec7f2",
+      "03ce61f1d155738a5e434fc8a61c3e104f891d1ec71576e8ad85abb68b34670d35",
+      "03ef2340518b5867b23598a9cf74611f8b98064f7d55cdb8c107c67b5efcbc5c77"
+		],
+		"numSignatures": 2,
+		"sender": "SM2R12RQCV9SCAZPM37VSCVP4X3EQK1Y70KCV7EDE"
+	}
+]
